Star lord

+++ app/Http/Controllers/ProductController.php
@@ -9,19 +9,39 @@ use App\Http\Requests;
 use \Cart as Cart;
 use Auth;
 use App\Order;
+use App\Item_info;
+use App\Mail\OrderMade;
+use Illuminate\Support\Facades\Mail;
+use Illuminate\Mail\Mailer;
+
 
 class ProductController extends Controller
 {
     public function getIndex()
     {
-        return view('shop.index');
+        $items = DB::select("SELECT * FROM desktops ORDER BY created_at DESC LIMIT 5");
+
+        $items2 = DB::select("SELECT * FROM laptops ORDER BY created_at DESC LIMIT 5");
+
+        $items3 = DB::select("SELECT * FROM accessories ORDER BY created_at DESC LIMIT 5");
+        return view('shop.index', ['items' => $items, 'items2' => $items2, 'items3' => $items3]);
     }
 
     public function showItem($id)
     {
         $table = $this->selectItemType($id);
         $item = DB::select("select * from " . $table . " where proid='" . $id . "'");
-        return view('shop.show', ['items' => $item]);
+
+        $specs = new Item_info();
+
+        $specs = $item[0]->itemDetails;
+        $specs = unserialize($specs);
+
+        $specs_arr = $specs->returnArr();
+
+        //dd($specs_arr['gui']);
+
+        return view('shop.show', ['items' => $item, 'specs' => $specs_arr]);
     }
 
     public function getCart()
@@ -64,11 +84,14 @@ class ProductController extends Controller
         }
     }
 
-    public function getAddToCart($id)
+    public function getAddToCart($id, Request $request)
     {
         $table = $this->selectItemType($id);
         $item = DB::select("select * from " . $table . " where proid='" . $id . "'");
         // $item = DB::select("select * from laptops where proid='LP001'");
+
+        $request->session()->flash('status', 'Task was successful!');
+
 
         foreach ($item as $itm) {
             $proid = $itm->proid;
@@ -112,21 +135,70 @@ class ProductController extends Controller
         $serializedContent = serialize($content);   //convert the object into a string
         $total = Cart::subtotal();
         //return response()->json(['returnURL' => $total], 200);
-        DB::insert('insert into orders (email,order_obj,total,delivery,paymentType) values (?,?,?,?,?)', [Auth::user()->email, $serializedContent, $total, $deliveryMethod, $paymentMethod]);
+        // DB::insert('insert into orders (email,order_obj,total,delivery,paymentType) values (?,?,?,?,?)', [Auth::user()->email, $serializedContent, $total, $deliveryMethod, $paymentMethod]);
 
-        Cart::destroy();
-        if ($paymentMethod == "pickup")    //pickup
-        {
-            return response()->json(['returnURL' => "www.google.lk"], 200);
-        } else if ($paymentMethod == "bank")    //pickup
-        {
-            return response()->json(['returnURL' => "bank baibee"], 200);
-        } else if ($paymentMethod == "paypal")    //pickup
-        {
-            return response()->json(['returnURL' => "paypal baibee"], 200);
+        $id = DB::table('orders')->insertGetId(
+            ['email' => Auth::user()->email, 'order_obj' => $serializedContent,
+                'total' => $total, 'delivery' => $deliveryMethod, 'paymentType' => $paymentMethod]
+        );
+
+//This code block crosscheck qty with available stock
+        $content = Cart::content();
+        $ptr = 0;
+        $outStock = array();
+        foreach ($content as $itm) {
+            $proid = $itm->id;
+            $qty = $itm->qty;
+
+            $curStock = DB::select("select stock from stock where proid='" . $proid . "'");
+            $curStock = $curStock[0]->stock;
+
+            if ($qty <= $curStock) {
+                continue;
+            } else {
+                array_push($outStock, $itm->name);
+                array_push($outStock, "dumber");
+                $ptr = 1;
+            }
+        }
+        if ($ptr == 1) {
+            return response()->json(['outstock' => $outStock], 200);
+        } else {
+            session(['orderID' => $id]);
+
+            Mail::to(Auth::user()->email)->send(New OrderMade());
+
+            Cart::destroy();
+            if ($paymentMethod == "pickup")    //pickup
+            {
+                \Session::put('pav_success', 'Payment success');
+                return response()->json(['returnURL' => "pay_later"], 200);
+            } else if ($paymentMethod == "bank")    //bank
+            {
+                return response()->json(['returnURL' => "paywithbank/$id"], 200);
+            } else if ($paymentMethod == "paypal")    //paypal
+            {
+                return response()->json(['returnURL' => "paywithpaypal/$total"], 200);
+            }
         }
 
-        //return redirect()->route('user.getCart');
+    }
+
+    public function getBank($id)
+    {
+        return view('shop.paywithbank', ['id' => $id]);
+    }
+
+    public function postBank(Request $request)
+    {
+        //dd($request);
+
+        DB::table('orders')
+            ->where('id', $request->id)
+            ->update(['payment_ref' => $request->ref]);
+
+        \Session::put('pav_success', 'Payment success');
+        return redirect()->route('user.getCart');
 
     }
